refactor from reaper to jobreaper
